# openapi.yaml
openapi: 3.0.3
info:
  title: 图书管理系统 API
  description: 基于Flask的图书管理系统RESTful API接口文档
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:5000
    description: 开发服务器
...
paths:
...
  /api/borrows:
    get:
      tags:
        - 借阅
      summary: 获取借阅记录列表
      description: 分页获取借阅记录（管理员可查看所有，普通用户只能查看自己的）
      security:
        - cookieAuth: []
      parameters:
        - name: page
          in: query
          description: 页码
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          description: 每页数量
          schema:
            type: integer
            default: 10
        - name: status
          in: query
          description: 借阅状态筛选
          schema:
            type: string
            enum: [borrowed, returned, ""]
      responses:
        '200':
          description: 成功获取借阅记录
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  records:
                    type: array
                    items:
                      $ref: '#/components/schemas/BorrowRecord'
                  total:
                    type: integer
                  pages:
                    type: integer
                  current_page:
                    type: integer

    post:
      tags:
        - 借阅
      summary: 办理借阅
      description: 为用户办理图书借阅（仅管理员）
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_id
                - book_id
              properties:
                user_id:
                  type: integer
                  description: 用户ID
                  example: 1
                book_id:
                  type: integer
                  description: 图书ID
                  example: 1
                days:
                  type: integer
                  description: 借阅天数
                  default: 30
                  example: 30
      responses:
        '201':
          description: 借阅成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 借阅成功
                  record:
                    $ref: '#/components/schemas/BorrowRecord'
        '400':
          description: 请求参数错误或图书无库存
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 需要管理员权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 用户或图书不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/borrows/{record_id}:
    get:
      tags:
        - 借阅
      summary: 获取借阅记录详情
      description: 根据ID获取借阅记录详情
      security:
        - cookieAuth: []
      parameters:
        - name: record_id
          in: path
          required: true
          description: 借阅记录ID
          schema:
            type: integer
      responses:
        '200':
          description: 成功获取借阅记录
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  record:
                    $ref: '#/components/schemas/BorrowRecord'
        '403':
          description: 无权查看此记录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 记录不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/borrows/{record_id}/return:
    put:
      tags:
        - 借阅
      summary: 办理归还
      description: 办理图书归还（仅管理员）
      security:
        - cookieAuth: []
      parameters:
        - name: record_id
          in: path
          required: true
          description: 借阅记录ID
          schema:
            type: integer
      responses:
        '200':
          description: 归还成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 归还成功
                  record:
                    $ref: '#/components/schemas/BorrowRecord'
        '400':
          description: 图书已归还
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 需要管理员权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 记录不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/borrows/user/{user_id}:
    get:
      tags:
        - 借阅
      summary: 查询用户借阅记录
      description: 获取指定用户的借阅记录
      security:
        - cookieAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          description: 用户ID
          schema:
            type: integer
        - name: page
          in: query
          description: 页码
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          description: 每页数量
          schema:
            type: integer
            default: 10
        - name: status
          in: query
          description: 借阅状态筛选
          schema:
            type: string
            enum: [borrowed, returned, ""]
      responses:
        '200':
          description: 成功获取借阅记录
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  records:
                    type: array
                    items:
                      $ref: '#/components/schemas/BorrowRecord'
                  total:
                    type: integer
                  pages:
                    type: integer
                  current_page:
                    type: integer
        '403':
          description: 无权查看此用户的借阅记录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/borrows/book/{book_id}:
    get:
      tags:
        - 借阅
      summary: 查询图书借阅记录
      description: 获取指定图书的借阅记录（仅管理员）
      security:
        - cookieAuth: []
      parameters:
        - name: book_id
          in: path
          required: true
          description: 图书ID
          schema:
            type: integer
        - name: page
          in: query
          description: 页码
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          description: 每页数量
          schema:
            type: integer
            default: 10
        - name: status
          in: query
          description: 借阅状态筛选
          schema:
            type: string
            enum: [borrowed, returned, ""]
      responses:
        '200':
          description: 成功获取借阅记录
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  records:
                    type: array
                    items:
                      $ref: '#/components/schemas/BorrowRecord'
                  total:
                    type: integer
                  pages:
                    type: integer
                  current_page:
                    type: integer
        '403':
          description: 需要管理员权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/borrows/overdue:
    get:
      tags:
        - 借阅
      summary: 获取逾期借阅记录
      description: 获取所有逾期未还的借阅记录（仅管理员）
      security:
        - cookieAuth: []
      parameters:
        - name: page
          in: query
          description: 页码
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          description: 每页数量
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: 成功获取逾期记录
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  records:
                    type: array
                    items:
                      $ref: '#/components/schemas/BorrowRecord'
                  total:
                    type: integer
                  pages:
                    type: integer
                  current_page:
                    type: integer
        '403':
          description: 需要管理员权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ==================== 用户接口 ====================
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          description: 用户ID
          example: 1
        username:
          type: string
          description: 用户名
          example: admin
        name:
          type: string
          description: 真实姓名
          example: 管理员
        phone:
          type: string
          description: 手机号/学号
          example: "10000000000"
        role:
          type: string
          description: 角色
          enum: [admin, user]
          example: admin
        created_at:
          type: string
          format: date-time
          description: 创建时间
          example: "2024-01-01T00:00:00"
    Book:
      type: object
      properties:
        id:
          type: integer
          description: 图书ID
          example: 1
        title:
          type: string
          description: 书名
          example: Python编程从入门到实践
        author:
          type: string
          description: 作者
          example: Eric Matthes
        isbn:
          type: string
          description: ISBN号
          example: "9787115428028"
        quantity:
          type: integer
          description: 总数量
          example: 5
        available:
          type: integer
          description: 可借数量
          example: 3
        is_available:
          type: boolean
          description: 是否可借
          example: true
        created_at:
          type: string
          format: date-time
          description: 创建时间
        updated_at:
          type: string
          format: date-time
          description: 更新时间
    BorrowRecord:
      type: object
      properties:
        id:
          type: integer
          description: 借阅记录ID
          example: 1
        user_id:
          type: integer
          description: 用户ID
          example: 1
        book_id:
          type: integer
          description: 图书ID
          example: 1
        user:
          $ref: '#/components/schemas/User'
        book:
          $ref: '#/components/schemas/Book'
        borrow_date:
          type: string
          format: date-time
          description: 借阅时间
          example: "2024-01-01T10:00:00"
        due_date:
          type: string
          format: date-time
          description: 应还时间
          example: "2024-01-31T10:00:00"
        return_date:
          type: string
          format: date-time
          nullable: true
          description: 实际归还时间
          example: null
        status:
          type: string
          description: 借阅状态
          enum: [borrowed, returned]
          example: borrowed
        is_overdue:
          type: boolean
          description: 是否逾期
          example: false
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          description: 错误信息
          example: 请求参数错误